overriding django-siwe nonce
Django siwe package does not pass the csrf token back in a response as it seems like it's intended for single domain usage.

Not being able to pass the csrf in a template tag means we have to get it from somewhere else, and the nonce request is the only non-authenticated view that we have.

I am not an expert in this field, so there is probably still some cleaning up of security, such as setting CSRF_TRUSTED_ORIGINS in the base/prod.py. But, that's a future item.