fix: Badge editing
Editing a badge should not allow the user to change delegates/managers in this form, as the setBadge method only allows adding delegates. Allowing them to edit in this form would be more confusing as they would expect removing from the list would remove, and would cost more gas by passing addresses that are already managers.