fix: ws connection
WS were all kinds of messed up for me. I believe this commit fixes it and everything currently works like a charm on localhost.

We were not accepting the connection prior to sending a connect message.

However, after diving in, I found that we cannot clear the sessionid from the client side as it is an HTTP_ONLY cookie and cannot be manipulated by the client. We can make an endpoint in the API that clears it, or probably better, have it clear on a /login/ call, but that's currently being inherited from django_siwe_auth.

We have to clear the sessionid in order to facilitate account switching.