fix: indexer websocket "event loop closed"
The indexer was broken due to (https://github.com/django/channels_redis/issues/332).

This resulted in the indexer constantly erroring out, and not having a way to recover because it was constantly stuck searching for the object in the database.

While there are still a few nuances with this, the API is accessible, is not blocked by the indexer, and "can" be independently scaled if that need arises; although is not implemented today.

There may still be a few bugs here, but we **cannot** move to version 4 of `channels-redis` otherwise everything will implode and leave you with no idea what is going on wrong (the package handles things on the backend that you don't think about and results in you debugging absolutely your entire project only to realize the library is unstable and usage should be delayed until stable v5 is released.)